Move managers/core to util/data_gatherer.
