+++ indra_db/managers/content_manager.py
@@ -26,7 +26,7 @@ from indra.util import UnicodeXMLTreeBuilder as UTB
 from indra_db.util import get_primary_db, get_db
 from indra_db.databases import texttypes, formats
 from indra_db.databases import sql_expressions as sql_exp
-from indra_db.managers.core import DataGatherer, DGContext
+from indra_db.util.data_gatherer import DataGatherer, DGContext
 
 
 try:
